Improve cacheing - only set from /jobs - only get from API route
