+++ package.json
@@ -24,6 +24,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.2.6",
+    "dotenv": "^10.0.0",
     "postcss": "^8.3.5",
     "tailwindcss": "^2.2.4"
   }
